fix(): 提取ConfigMonitor接口
