+++ apollo-client/src/main/java/com/ctrip/framework/apollo/internals/DefaultInjector.java
@@ -104,7 +104,7 @@ public class DefaultInjector implements Injector {
       bind(RemoteConfigLongPollService.class).in(Singleton.class);
       bind(YamlParser.class).in(Singleton.class);
       bind(PropertiesFactory.class).to(DefaultPropertiesFactory.class).in(Singleton.class);
-      bind(ConfigMonitor.class).in(Singleton.class);
+      bind(ConfigMonitor.class).to(DefaultConfigMonitorImpl.class).in(Singleton.class);
     }
   }
 }
