+++ apollo-core/src/main/java/com/ctrip/framework/apollo/metrics/Metrics.java
@@ -23,9 +23,6 @@ public abstract class Metrics {
     }
   }
 
-//  public static GaugeMetricsSample<?> buildGaugeMetricsSample(String name, Object value) {
-//    return GaugeMetricsSample.builder().name(name).value(value).build();
-//  }
 
   public static void push(MetricsEvent event) {
     if (isMetricsEnabled()) {
